+++ MiDAS/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.0.9008 - 11/05/2019
++ introduces countsToHlaCalls that allows converting HLA counts table back to
+  HLA calls data frame under additive inheritance model 
+
 ## 0.0.0.9007 - 8/05/2019
 + forwardConditionalSelection has been renamed to analyzeConditionalAssociations,
   now it returns tibble containing results from conditional testing as the new